Add backpacks to inventories
